requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
distlib==0.3.6
Django==4.2.4
dnspython==2.4.2
filelock==3.12.2
Flask==2.3.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
platformdirs==3.5.3
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
speedtest==0.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
validators==0.22.0
virtualenv==20.23.1
Werkzeug==2.3.7
WTForms==3.0.1
